requirements.txt
anyio==4.8.0
asgiref==3.8.1
certifi==2024.12.14
Django==5.1.4
django-admin-sortable2==2.2.4
djangorestframework==3.15.2
et_xmlfile==2.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
openpyxl==3.1.5
python-telegram-bot==21.10
sniffio==1.3.1
sqlparse==0.5.3
tzdata==2024.2
